requirements.txt
keyboard==0.13.5
mss==10.0.0
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
psutil==6.1.0
